+++ packages/networks/ergo-explorer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosen-chains/ergo-explorer-network
 
+## 10.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-chains/abstract-chain@16.0.1
+  - @rosen-chains/ergo@14.1.0
+  - @rosen-clients/ergo-explorer@2.1.3
+
 ## 10.0.3
 
 ### Patch Changes